split side navigation and top row components, got basic layout working
