--- frontend/src/components/content/Content.module.scss
@@ -1,9 +1,0 @@
-@import 'styles/spacings';
-@import 'styles/colors';
-
-.contentContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  flex: 1;
-}
