+++ frontend/src/components/layout/Layout.tsx
@@ -1,14 +1,14 @@
 import React, { FC, FunctionComponent, useCallback, useState } from "react";
 import classes from "./Layout.module.scss";
-import Navigation from "../navigation/Navigation";
-import Content from "../content/Content";
+import { TopRow } from "../top-row/TopRow";
 import Logs from "../../pages/logs/Logs";
 import { useLogDrawer } from "../../hooks/use-log-drawer";
 import { SideBar } from "components/sidebar/SideBar";
 import { Route, RouteProps, useHistory } from "react-router-dom";
 import { ReactComponent as IconBackButton } from "../../assets/icons/back-button.svg";
 import classNames from "classnames";
-import { Breadcrumbs } from "../navigation/Breadcrumbs";
+import { Breadcrumbs } from "../breadcrumbs/Breadcrumbs";
+import { SideNavigation } from "../side-navigation/SideNavigation";
 
 export const RouteWithBackButton: FC<RouteProps> = (props) => {
   const history = useHistory();
@@ -48,12 +48,15 @@ const Layout: FunctionComponent = ({ children }) => {
   }, [logDrawerSize]);
 
   return (
-    <div className={classes.layoutContainer}>
-      <Navigation isSidebarOpen={isSidebarOpen} toggleSidebar={toggleSidebar} />
-      <Breadcrumbs />
-      <SideBar toggled={isSidebarOpen} toggleSidebar={toggleSidebar} />
-      <Content className={classes.content}>{children} </Content>
-      <Logs className={classNames(classes.logs, getLogDrawerLayoutClass())} />
+    <div className={classes.root}>
+      <SideNavigation />
+      <div className={classes.mainContent}>
+        <TopRow isSidebarOpen={isSidebarOpen} toggleSidebar={toggleSidebar} />
+        <Breadcrumbs />
+        <SideBar toggled={isSidebarOpen} toggleSidebar={toggleSidebar} />
+        <div className={classes.body}>{children} </div>
+        <Logs className={classNames(classes.logs, getLogDrawerLayoutClass())} />
+      </div>
     </div>
   );
 };
