--- frontend/src/components/navigation/NavigationItem.module.scss
@@ -1,66 +1,0 @@
-@import "styles/colors";
-@import "styles/typography";
-@import "styles/spacings";
-
-.root {
-  width: 120px;
-  .navLink {
-    display: flex;
-    color: $light-grey;
-    font-size: 15px;
-    text-decoration: none;
-    align-items: center;
-    flex-direction: column;
-    gap: 6px;
-    line-height: 19px;
-    padding: 11px 45px 14px 45px;
-    border-right: 1px solid $background;
-
-    @media (max-width: 1500px) {
-      padding: 11px 23px 14px 23px;
-    }
-
-    &:hover {
-      opacity: 0.9;
-    }
-
-    &.active {
-      color: $yellow;
-      background: #4e516b;
-    }
-
-    & > div:first-child {
-      margin-right: $spacing-base;
-    }
-
-    .counters {
-      display: flex;
-      flex-direction: row;
-      justify-content: center;
-      gap: 5px;
-
-      > * {
-        font-size: $font-size-xsmall;
-        line-height: calc(#{$font-size-xxsmall} + 2px);
-        color: $color-main-text;
-        border-radius: 8px;
-        text-align: center;
-        font-weight: 400;
-        padding: 2px 5px;
-      }
-      .totalCounter {
-        background-color: $grey;
-      }
-
-      .unreadCounter {
-        background-color: $blue;
-      }
-    }
-
-    &.active {
-      .totalCounter {
-        background-color: $yellow;
-      }
-    }
-  }
-}
