+++ frontend/src/components/side-navigation/SideNavigation.module.scss
@@ -1,0 +1,4 @@
+.root {
+  display: flex;
+  flex-direction: column;
+}
